Added the recipe for saddles. Oh.
The new items have no textures. Please excuse that. I have commented out
the code for the Ruby Block Recipe as it was causing me to crash. If you
know what is going on please tell me. Oh, and if someone could be nice
enough to volunteer every once in a while to do some textures for me. If
smeone could do the textures fr these three items i have added. Stirips,
Double Stirips and Buckles i would greatly appreciate it. As always,
feel free to make changes to my code if you feel that it is nessesary. I
will have a website up soon to put more info on the mod.